requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cfgv==3.4.0
colorlog==6.9.0
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
openai==1.68.2
packaging==24.2
pandas==2.2.3
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
ruff==0.11.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
virtualenv==20.29.3
